project/webapp_forGit copy/requirements.txt
Flask==2.3.2
firebase-admin==6.0.1
google-cloud-storage==2.9.0
